requirements.txt
Django>=5.1.1
django-ninja>=1.1.0
PyJWT>=2.8.0
django-cors-headers>=4.3.0
gunicorn>=23.0.0
pandas>=2.0.0
openpyxl>=3.1.0